refactor(#142/*): reservation success code global/response 로 이동
