+++ src/main/java/org/sopt/snappinserver/api/v1/reservation/controller/ReservationController.java
@@ -1,7 +1,7 @@
 package org.sopt.snappinserver.api.v1.reservation.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.snappinserver.api.v1.reservation.code.ReservationSuccessCode;
+import org.sopt.snappinserver.global.response.code.reservation.ReservationSuccessCode;
 import org.sopt.snappinserver.api.v1.reservation.dto.request.CreateReservationReviewRequest;
 import org.sopt.snappinserver.api.v1.reservation.dto.request.RequestPaymentReservationRequest;
 import org.sopt.snappinserver.api.v1.reservation.dto.response.CancelReservationResponse;
