+++ src/main/java/org/sopt/snappinserver/global/response/code/reservation/ReservationSuccessCode.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.api.v1.reservation.code;
+package org.sopt.snappinserver.global.response.code.reservation;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
